+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -2,7 +2,4 @@
 
 /// Bytecode unroller
 pub mod bytecode_unroller;
-/// Bytecode circuit tester
-#[cfg(any(feature = "test", test))]
-pub mod dev;
 pub(crate) mod param;
