--- zkevm-circuits/src/bytecode_circuit/dev.rs
@@ -1,95 +1,0 @@
-use super::bytecode_unroller::{
-    unroll, BytecodeCircuit, BytecodeCircuitConfig, BytecodeCircuitConfigArgs, UnrolledBytecode,
-};
-use crate::table::{BytecodeTable, KeccakTable};
-use crate::util::{Challenges, SubCircuit, SubCircuitConfig};
-use eth_types::Field;
-use halo2_proofs::{
-    circuit::Layouter,
-    plonk::{ConstraintSystem, Error},
-};
-use halo2_proofs::{circuit::SimpleFloorPlanner, dev::MockProver, plonk::Circuit};
-
-impl<F: Field> Circuit<F> for BytecodeCircuit<F> {
-    type Config = (BytecodeCircuitConfig<F>, Challenges);
-    type FloorPlanner = SimpleFloorPlanner;
-
-    fn without_witnesses(&self) -> Self {
-        Self::default()
-    }
-
-    fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
-        let bytecode_table = BytecodeTable::construct(meta);
-        let keccak_table = KeccakTable::construct(meta);
-        let challenges = Challenges::construct(meta);
-
-        let config = {
-            let challenges = challenges.exprs(meta);
-            BytecodeCircuitConfig::new(
-                meta,
-                BytecodeCircuitConfigArgs {
-                    bytecode_table,
-                    keccak_table,
-                    challenges,
-                },
-            )
-        };
-
-        (config, challenges)
-    }
-
-    fn synthesize(
-        &self,
-        (config, challenges): Self::Config,
-        mut layouter: impl Layouter<F>,
-    ) -> Result<(), Error> {
-        let challenges = challenges.values(&mut layouter);
-
-        config.keccak_table.dev_load(
-            &mut layouter,
-            self.bytecodes.iter().map(|b| &b.bytes),
-            &challenges,
-        )?;
-        self.synthesize_sub(&config, &challenges, &mut layouter)?;
-        Ok(())
-    }
-}
-
-impl<F: Field> BytecodeCircuit<F> {
-    /// Verify that the selected bytecode fulfills the circuit
-    pub fn verify_raw(k: u32, bytecodes: Vec<Vec<u8>>) {
-        let unrolled: Vec<_> = bytecodes.iter().map(|b| unroll(b.clone())).collect();
-        Self::verify(k, unrolled, true);
-    }
-
-    pub(crate) fn verify(k: u32, bytecodes: Vec<UnrolledBytecode<F>>, success: bool) {
-        let circuit = BytecodeCircuit::<F>::new(bytecodes, 2usize.pow(k));
-
-        let prover = MockProver::<F>::run(k, &circuit, Vec::new()).unwrap();
-        let result = prover.verify();
-        if let Err(failures) = &result {
-            for failure in failures.iter() {
-                println!("{}", failure);
-            }
-        }
-        assert_eq!(result.is_ok(), success);
-    }
-}
-
-/// Test bytecode circuit with unrolled bytecode
-pub fn test_bytecode_circuit_unrolled<F: Field>(
-    k: u32,
-    bytecodes: Vec<UnrolledBytecode<F>>,
-    success: bool,
-) {
-    let circuit = BytecodeCircuit::<F>::new(bytecodes, 2usize.pow(k));
-
-    let prover = MockProver::<F>::run(k, &circuit, Vec::new()).unwrap();
-    let result = prover.verify_par();
-    if let Err(failures) = &result {
-        for failure in failures.iter() {
-            println!("{}", failure);
-        }
-    }
-    assert_eq!(result.is_ok(), success);
-}
